Распределена ответственность в классе Token.
Сделана адаптация под новую сигнатуру класса TagInfo
